This is a try to get real 16bit data from PNG. Problem: simplified API of libPNG pre-multiply color in 16 bits mode ! Solution: We must use low level libPNG API Advantage: Better memory usage (no more big buffer copy) Drawbacks: More difficult to do (yes, but this is exactly why I'm coding ImaGL)

+++ imaGL/imaGL.cpp
@@ -115,10 +115,10 @@ namespace ImaGL {
   }
 
   const unsigned char* CImaGL::pixels()      const { return m_pData->m_vRawData.data(); }
-  int                   CImaGL::width()       const { return m_pData->m_nWidth; }
-  int                   CImaGL::height()      const { return m_pData->m_nHeight; }
-  CImaGL::EPixelFormat  CImaGL::pixelformat() const { return m_pData->m_PixelFormat; }
-  CImaGL::EPixelType    CImaGL::pixeltype()   const { return m_pData->m_PixelType; }
-  int                   CImaGL::pixelsize()   const { return m_nPixelSize; }
+  size_t               CImaGL::width()       const { return m_pData->m_nWidth; }
+  size_t               CImaGL::height()      const { return m_pData->m_nHeight; }
+  CImaGL::EPixelFormat CImaGL::pixelformat() const { return m_pData->m_PixelFormat; }
+  CImaGL::EPixelType   CImaGL::pixeltype()   const { return m_pData->m_PixelType; }
+  size_t               CImaGL::pixelsize()   const { return m_nPixelSize; }
 
 }
